# apps/corpus/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin corpus
  version: "1.0.0"
  description: Per-DID corpus indexing and BM25 search service — ingests threads from local workspace files and GitHub repositories, and serves evidence-based search over them.
servers:
  - url: http://localhost:8003
    description: internal (not publicly routed; consumed only by the kernel MCP proxy)

tags:
  - name: corpus

components:
  securitySchemes:
    kernelProxyOnly:
      type: apiKey
      in: header
      name: (none)
      description: >-
        This service accepts no credential of its own. Every /corpus/:did/*
        route trusts its :did path parameter as given — auth is enforced one
        layer up, by the kernel MCP proxy (apps/kernel/src/lib/mcp/tools/corpus.ts),
        which requires an active corpus:read/corpus:write grant before it
        forwards a call here over the internal network.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }

paths:
  /health:
    get:
      tags: [corpus]
      operationId: health
      summary: Liveness check for the corpus service.
      description: Public and unauthenticated by design — used for process/container health probes only.
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  service: { type: string }

  /spec:
    get:
      tags: [corpus]
      operationId: getSpec
      summary: Serve this OpenAPI spec (YAML) so discovery tooling can enumerate corpus's routes.
      description: Public and unauthenticated by design, matching every sibling service's /spec (or /{service}/api/spec) route.
      responses:
        "200":
          description: OpenAPI YAML spec
          content:
            text/yaml:
              schema: { type: string }

  /corpus/{did}/status:
    get:
      tags: [corpus]
      operationId: corpusStatus
      summary: Report indexed sources, thread counts, and freshness for a DID's corpus.
      description: No auth enforced at this layer — see kernelProxyOnly. Today this route answers for any :did without a scope check.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Corpus status
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources: { type: array, items: { type: object } }
                  threadCount: { type: integer }

  /corpus/{did}/search:
    post:
      tags: [corpus]
      operationId: corpusSearch
      summary: Search a DID's corpus for threads matching a query, returning scored evidence excerpts.
      description: No auth enforced at this layer — proxied only via the kernel MCP corpus_search tool, which requires corpus:read.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                sourceType: { type: string }
                source: { type: string }
                state: { type: string }
                type: { type: string }
                labels: { type: array, items: { type: string } }
                author: { type: string }
                limit: { type: integer }
                budget: { type: integer }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/ingest:
    post:
      tags: [corpus]
      operationId: corpusIngest
      summary: Ingest raw ThreadDocument[] directly, or crawl+ingest a `{ source }` (local:workspace or github:owner/repo).
      description: No auth enforced at this layer — proxied only via the kernel MCP corpus_load tool, which requires corpus:write.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
                  required: [source]
                  properties:
                    source: { type: string }
      responses:
        "200":
          description: Ingest result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/sources:
    post:
      tags: [corpus]
      operationId: corpusRegisterSource
      summary: Register a source (`{ source, type }`) and perform its initial full crawl — the GitHub adapter's entry point (#1729).
      description: No auth enforced at this layer — not yet proxied by a kernel MCP tool; called directly today.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string, description: "e.g. \"local:workspace\" or \"github:owner/repo\"" }
                type: { type: string, description: "Optional; must match the source's own prefix, e.g. \"github\"." }
      responses:
        "200":
          description: Registration + initial crawl result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/sync:
    post:
      tags: [corpus]
      operationId: corpusSync
      summary: Incrementally sync a previously-registered `{ source, cursor }` (local:workspace or github:owner/repo).
      description: No auth enforced at this layer — proxied only via the kernel MCP corpus_sync tool, which requires corpus:write.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string }
                cursor: { type: string, nullable: true }
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
                  cursor: { type: string, nullable: true }
                  hasMore: { type: boolean }
        "501":
          description: Source is unset or not a recognized (workspace/github) source
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/crawl:
    post:
      tags: [corpus]
      operationId: corpusCrawl
      summary: Full crawl+ingest of a `{ source }` (local:workspace or github:owner/repo), without registering it.
      description: No auth enforced at this layer — not yet proxied by a kernel MCP tool; called directly today.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string }
      responses:
        "200":
          description: Crawl result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/source:
    delete:
      tags: [corpus]
      operationId: corpusDeleteSource
      summary: Delete every indexed thread for a `{ source }` from a DID's corpus.
      description: No auth enforced at this layer — not yet proxied by a kernel MCP tool; called directly today.
      security: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string }
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }